backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Authentication & Security
pyjwt==2.10.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# HTTP Clients
httpx==0.28.1
requests==2.32.3

# Image Processing (for local file storage)
Pillow==11.0.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==33.1.0

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2

# Logging
python-json-logger==3.2.1

# Development
ipython==8.31.0
watchfiles==1.0.3
